requirements.txt
streamlit
openai-whisper
deep-translator
ffmpeg-python
torch
setuptools
